lsp-devtools: restructure code, start thinking about config
