--- lib/lsp-devtools/lsp_devtools/client/__init__.py
@@ -1,3 +1,0 @@
-from .lsp_client import LanguageClient
-
-__all__ = ("LanguageClient",)
