+++ lib/lsp-devtools/lsp_devtools/inspector/__init__.py
@@ -1,0 +1,3 @@
+from .message_browser import MessageBrowser
+
+__all__ = ("MessageBrowser",)
